feat(markdown): 新增 AI 流式智能对话助手并完成表格渲染根因治理
1. 新增 AiChatActivity 实战页面：支持多轮对话、RecyclerView Payload 局部增量刷新、手势感知与悬浮回底按钮；
2. 根因重构 StreamTableRowSpan：首轮测量期 (getSize) 即时构建 StaticLayout 精准分配行高，彻底消除行间压叠与单元格压线；
3. 优化 MarkdownStreamFixer：隔离 GFM 表格语法行，修复由于光标插入破坏 AST 导致的闪烁问题；
4. 沉淀治理铁律：在 AGENTS.md 与 docs/conventions.md 中固化「从根因治理，严禁补丁式修改」的不变量。

+++ modules/module_markdown/src/main/java/com/example/william/my/module/markdown/activity/MarkwonHighlightActivity.kt
@@ -65,10 +65,10 @@ class MarkwonHighlightActivity : BasicLayoutActivity() {
 
         val tableTheme = TableTheme.Builder()
             .tableBorderWidth(dpToPx(1))
-            .tableBorderColor(0x33888888.toInt())
+            .tableBorderColor(0x33888888)
             .tableCellPadding(dpToPx(8))
-            .tableHeaderRowBackgroundColor(0x18888888.toInt())
-            .tableEvenRowBackgroundColor(0x08888888.toInt())
+            .tableHeaderRowBackgroundColor(0x18888888)
+            .tableEvenRowBackgroundColor(0x08888888)
             .build()
 
         // 1. Darkula 暗黑代码主题
